added sectors

+++ src/diffusionlimitedaggregation/Point.java
@@ -1,17 +1,17 @@
 package diffusionlimitedaggregation;
 
 public class Point {
-    private int x,y;
-    public Point(int x, int y){
+    private double x,y;
+    public Point(double x, double y){
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
